fix: refactored openRouter feature based on pull request reviews
