+++ src/components/Message/AppMessage/Instructions.tsx
@@ -16,7 +16,6 @@ import RevealablePasswordInput from "../../RevealablePasswordInput";
 import { useSettings } from "../../../hooks/use-settings";
 import { openRouterPkceWindow, validateOpenAiApiKey } from "../../../lib/ai";
 import { OPENAI_API_URL, OPENROUTER_API_URL } from "../../../lib/settings";
-import useCodeExchange from "../../../hooks/use-codeExchange";
 
 const ApiKeyInstructionsText = `Welcome to ChatCraft, a developer-focused AI assistant. I can help you write code, visualize it with mermaid, html and even run it. You can further refine code by editig, deleting and retrying model responses.
 
@@ -39,15 +38,14 @@ function Instructions(props: MessageBaseProps) {
   const { settings, setSettings } = useSettings();
   const [isValidating, setIsValidating] = useState(false);
   const [isInvalid, setIsInvalid] = useState(false);
-  const { apiKeyValue, handleSetApiKey } = useCodeExchange();
 
   // Override the text of the message
   const message = new ChatCraftAppMessage({ ...props.message, text: ApiKeyInstructionsText });
 
   const handleApiKeySubmit = (e: FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-
-    const apiKey = apiKeyValue;
+    const data = new FormData(e.target as HTMLFormElement);
+    const apiKey = data.get("openai-api-key");
     if (typeof apiKey !== "string") {
       return;
     }
@@ -95,8 +93,6 @@ function Instructions(props: MessageBaseProps) {
                 type="password"
                 name="openai-api-key"
                 bg="white"
-                onChange={(e) => handleSetApiKey(e.target.value)}
-                value={apiKeyValue}
                 _dark={{ bg: "gray.700" }}
                 required
               />
